Fix deadlocks of bitmap
Now, we have the two type of deadlocks of bitmap.

One is bitmap->btree->lock.

    write bitmap buffer
        map_region(bitmap)
            down_write(bitmap->btree->lock) <- this
	        balloc()
		    blockread()
		        map_region(bitmap)
		            down_read(bitmap->btree->lock) <- this

On the above path, it will takes write-lock and read-lock, so it is
deadlock.

To solve this, map_region() remembers whether it already taked
down_write() or not (get_bitmap_write()). And if it has down_write(),
second map_region() doesn't take down_read().

The another one is bitmap->i_mutex

        task 0                                 task 1
    down_write(bitmap->btree->lock)
                                       balloc()
					 bitmap->i_mutex
					   blockread()
		                             map_region(bitmap)
		                               down_read(bitmap->btree->lock)
        balloc()
	    bitmap->i_mutex

So, this patch is AB BA deadlock.

To solve this, this doesn't take bitmap->i_mutex durning
blockread(). This means it protects one block only, so if balloc() was
failed, it doesn't mean 0 free blocks.

The another task can change other bitmap buffers while checking one
buffer. However, this shouldn't be problem, because it can be happen
even if protect all buffers on user's view. Because even if user was
checking free blocks, another tasks still can use after checking.

The situation of locking may be changed with atomic commit, we should
revisit locking after/with atomic commit.