Cleanup/Fix the logging of bitmap
To clear, this adds "b" prefix to alloc/free.

And this separates BFREE and BFREE_ON_FLUSH (the name may not be
good). BFREE_ON_FLUSH is logging for sb->deflush, not sb->defree.

Then, this fixes log_bfree() in map_region(). It should adds deferred
bfree entry, and uses new BFREE_ON_FLUSH for bitmap.