Add find_dirty_inode(), and merge open_inode() to iget()
The inodes can be on sb->dirty_inodes after iput(). Right now, we
don't need this. However, if we open the inode, we have to find the
inode on sb->dirty_inodes. (NOTE, this is not finding all in-core
inodes. If we want to reopen the inode, we would need to find all
in-core inodes)

So, this adds the find_dirty_inode() to iget(). And iget() finds dirty
inode, and if not found, allocate inode and open_inode().