Defer to add root bnode to data btree for atomic-commit
This defers to add root bnode to data btree until actual request
(map_region).

This implementation is not using variable inode size ability.
I.e. this writes the root/depth (0, 0) even if btree doesn't have
root bnode.

With this change, map_region() doesn't return 0 when btree->root.depth == 0.
(In userspace, it was true for bitmap on make_tux3(). And in kernel,
sb->logmap.) Instead, this handles it as hole, this would be ok,
however may need to rethink.

NOTE: this doesn't delete the empty btree with truncate.