Fix reserve_superblock() in make_tux3()
reserve_superblock() is a bit tricky. It wants to reserve the blocks
early before being grabbed by others.  But allocating blocks for this
delta must be after sb->delta (or ->rollup) increment. Otherwise,
rollup and delta may write the both of bitmap data and balloc log in
this commit.

So, this provide the callback to do it.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>